Updated package to stop using deprecated methods in github.com/mrjones/oauth
Before, this package would use the oauth.Consumer.Get method for sending
authenticated requests to the Yahoo API. The updates to the oauth
package now makes it easier to create a http.Client that supports
authenticated requests.

As a result, this package as been updated to use a http.Client (passed
in as goff.HTTPClient) instead of the oauth.Consumer. Although in most
cases it will still be easier to use the github.com/mrjones/oauth
package to authenticate with the service, it is no longer required.

API Changes:

- Removed NewOAuthClient and NewCachedOAuthClient functions
- Remove OAuthConsumer interface
- Added NewClient and NewCachedClient
- Added HTTPClient interface